more use of 1L etc
git-svn-id: https://svn.r-project.org/R/trunk@47197 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/base/R/aperm.R
@@ -17,6 +17,6 @@
 aperm <- function(a, perm, resize=TRUE)
 {
     if (missing(perm))
-	perm <- integer(0) # will reverse the order
+	perm <- integer(0L) # will reverse the order
     .Internal(aperm(a, perm, resize))
 }
