+++ src/library/base/R/sort.R
@@ -106,10 +106,10 @@ order <- function(..., na.last = TRUE, decreasing = FALSE)
     if(any(diff(sapply(z, length)) != 0))
         stop("argument lengths differ")
     ans <- sapply(z, is.na)
-    if(is.list(ans)) return(integer(0)) # happens for 0-length input
+    if(is.list(ans)) return(integer(0L)) # happens for 0-length input
     ok <- if(is.matrix(ans)) !apply(ans, 1, any) else !any(ans)
-    if(all(!ok)) return(integer(0))
-    z[[1]][!ok] <- NA
+    if(all(!ok)) return(integer(0L))
+    z[[1L]][!ok] <- NA
     ans <- do.call("order", c(z, decreasing=decreasing))
     keep <- seq_along(ok)[ok]
     ans[ans %in% keep]
@@ -159,5 +159,5 @@ xtfrm.Surv <- function(x)
 .gtn <- function(x, strictly)
 {
     n <- length(x)
-    if(strictly) all(x[-1] > x[-n]) else all(x[-1] >= x[-n])
+    if(strictly) all(x[-1L] > x[-n]) else all(x[-1L] >= x[-n])
 }
